refactor(core): split LLM/embedding configs into types::llm submodule
Continues the `core/src/types.rs` kitchen-sink split. Moves the entire
LLM + embedding configuration block into `assistant_core::types::llm::*`:

- `LlmProviderKind`, `EmbeddingProviderKind`
- `EmbeddingConfig`, `LlmConfig`
- Anthropic / OpenAI / Moonshot / OpenRouter option sub-structs
- `default_llm_*` / `default_embedding_*` helpers

`types.rs` adds `use self::llm::LlmConfig;` for the `AssistantConfig`
field, exposes `pub(crate) fn default_true()` so the submodule can share
it via `super::default_true`, and updates the test module to import the
moved types.

Call-site changes (8 files):
- `crates/core/src/config.rs` (test refs)
- `crates/llm-provider/src/{lib,anthropic/mod,moonshot/mod,ollama/{mod,client},openrouter/mod,openai/provider}.rs`
- `crates/interface-cli/src/{main,cmd_doctor}.rs`
- `crates/web-ui/src/lib.rs`
