+++ crates/interface-cli/src/cmd_doctor.rs
@@ -8,9 +8,10 @@ use std::path::{Path, PathBuf};
 
 use anyhow::Result;
 
+use assistant_core::AssistantConfig;
+use assistant_core::types::llm::LlmProviderKind;
 use assistant_core::types::skills_mcp::McpTransportConfig;
 use assistant_core::types::storage::BusKind;
-use assistant_core::{AssistantConfig, LlmProviderKind};
 
 // ── Helpers ──────────────────────────────────────────────────────────────────
